added QP/C++-ThreadX port, updated uC/OS-II port

+++ examples/qt/mingw/dpp-gui/main.cpp
@@ -1,13 +1,13 @@
 //****************************************************************************
 // Product: QP/C++ example for Qt5
-// Last Updated for Version: QP 5.1.1/Qt 5.1.1
-// Date of the Last Update:  Nov 05, 2013
+// Last Updated for Version: QP 5.3.0/Qt 5.1.1
+// Last updated on  2014-04-21
 //
 //                    Q u a n t u m     L e a P s
 //                    ---------------------------
 //                    innovating embedded systems
 //
-// Copyright (C) 2002-2013 Quantum Leaps, LLC. All rights reserved.
+// Copyright (C) Quantum Leaps, www.state-machine.com.
 //
 // This program is open source software: you can redistribute it and/or
 // modify it under the terms of the GNU General Public License as published
@@ -28,9 +28,8 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 //
 // Contact information:
-// Quantum Leaps Web sites: http://www.quantum-leaps.com
-//                          http://www.state-machine.com
-// e-mail:                  info@quantum-leaps.com
+// Web:   www.state-machine.com
+// Email: info@state-machine.com
 //****************************************************************************
 #include "gui.h"
 //-----------------
@@ -43,7 +42,7 @@ static QP::QEvt const *l_philoQueueSto[N_PHILO][N_PHILO];
 static QP::QSubscrList l_subscrSto[DPP::MAX_PUB_SIG];
 
 // storage for event pools...
-static QF_MPOOL_EL(DPP::TableEvt) l_smlPoolSto[2*N_PHILO];       // small pool
+static QF_MPOOL_EL(DPP::TableEvt) l_smlPoolSto[2*N_PHILO]; // small pool
 
 //............................................................................
 int main(int argc, char *argv[]) {
@@ -52,27 +51,27 @@ int main(int argc, char *argv[]) {
 
     gui.show();
 
-    QP::QF::init();                                // initialize the framework
-    BSP_init();                                          // initialize the BSP
+    QP::QF::init(); // initialize the framework
+    BSP_init();     // initialize the BSP
 
-                                                     // object dictionaries...
+    // object dictionaries...
     QS_OBJ_DICTIONARY(l_smlPoolSto);
 
     QP::QF::psInit(l_subscrSto, Q_DIM(l_subscrSto)); // init publish-subscribe
 
-                                                  // initialize event pools...
+    // initialize event pools...
     QP::QF::poolInit(l_smlPoolSto,
                      sizeof(l_smlPoolSto), sizeof(l_smlPoolSto[0]));
 
-                                                // start the active objects...
-    for (uint8_t n = 0U; n < N_PHILO; ++n) {
-        DPP::AO_Philo[n]->start((uint8_t)(n + 1),
+    // start the active objects...
+    for (uint_fast8_t n = 0U; n < N_PHILO; ++n) {
+        DPP::AO_Philo[n]->start((uint_fast8_t)(n + 1),
                                 l_philoQueueSto[n], Q_DIM(l_philoQueueSto[n]),
                                 (void *)0, 1024U*4U);
     }
-    DPP::AO_Table->start((uint8_t)(N_PHILO + 1),
+    DPP::AO_Table->start((uint_fast8_t)(N_PHILO + 1),
                            (QP::QEvt const **)0, (uint32_t)0,
-                           (void *)0, (uint32_t)0);
+                           (void *)0, (uint_fast16_t)0);
 
-    return QP::QF::run();                                // calls qApp->exec()
+    return QP::QF::run(); // calls qApp->exec()
 }
